% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{achv_color_map}
\alias{achv_color_map}
\title{OHA Achievement Colors}
\format{
A data frame with three variables and 4 rows
\describe{
  \item{achv_desc}{description of each achievement level}
  \item{achv_color}{hex code of the associated color to the achievement level}
  \item{achv_color_classic}{classic/original hex code of the associated color to the achievement level}
}
}
\usage{
data(achv_color_map)
}
\description{
A tibble of the OHA Colors associated with different target achievement
thresholds - 'At Risk', 'Concerned', 'On Target', and 'Above Target'. This
table includes  the current mapping as well as the original/classic one.
These colors are used in association with `adorn_achievement`.
}
\keyword{datasets}
