% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_funding_type.R
\name{apply_funding_type}
\alias{apply_funding_type}
\title{Apply Funding Type}
\usage{
apply_funding_type(df)
}
\arguments{
\item{df}{Financial Structured Dataset data frame}
}
\value{
a new column, funding_type
}
\description{
When working with financial or HRH data, its often useful to determine
whether funding or staffing are counted as "service delivery" or "non-service
delivery"
}
