% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cascade_ind}
\alias{cascade_ind}
\title{MER Clinical Cascade indicators}
\format{
A list of clinical cascade indicators
\describe{
  \item{cascade_ind}{indicator names}
}
}
\usage{
data(cascade_ind)
}
\description{
List of indicators in the MER that comprise the clinical cascade. These
indicators are needed for calculating the 90s/95s plus linkage and viral load
coverage.
}
\keyword{datasets}
