% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{convert_names}
\alias{convert_names}
\title{Convert variables for alignment}
\usage{
convert_names(df)
}
\arguments{
\item{df}{data frame from read_psd()}
}
\description{
Convert variables for alignment
}
\keyword{internal}
