% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{convert_oldformat}
\alias{convert_oldformat}
\title{Convert any old PSDs to new format}
\usage{
convert_oldformat(df)
}
\arguments{
\item{df}{data frame from read_psd()}
}
\description{
Convert any old PSDs to new format
}
\keyword{internal}
