% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_official.R
\name{extract_datim_names}
\alias{extract_datim_names}
\title{Extract Data from DATIM}
\usage{
extract_datim_names(ou, username, password)
}
\arguments{
\item{url}{url for query}
}
\value{
converts json to data frame
}
\description{
Extract Data from DATIM
}
\keyword{internal}
