% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract MSD Source Information}
\usage{
extract_metadata(path, type)
}
\arguments{
\item{path}{path to the folder containing MSDs or specific MSD file}

\item{type}{not required unless providing a folder in `path`;
default = "OU_IM_FY2*", other examples include: "PSNU_IM", "NAT_SUBNAT",
"PSNU", "Financial", "HRH"}
}
\value{
dataframe of information related to what is being asked in `return`
}
\description{
Extract MSD Source Information
}
\seealso{
Other metadata: 
\code{\link{get_metadata}()},
\code{\link{source_info}()}
}
\concept{metadata}
\keyword{internal}
