% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\name{extract_path_local}
\alias{extract_path_local}
\title{Extract MSD path locally}
\usage{
extract_path_local(path, type)
}
\arguments{
\item{path}{path to the folder containing MSDs or specific MSD file}

\item{type}{not required unless providing a folder in `path`;
default = "OU_IM_FY2*", other examples include: "PSNU_IM", "NAT_SUBNAT",
"PSNU", "Financial", "HRH"}
}
\description{
Extract MSD path locally
}
\keyword{internal}
