% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\name{extract_path_s3}
\alias{extract_path_s3}
\title{Extract MSD path on s3 from Data Mart}
\usage{
extract_path_s3(path, type)
}
\arguments{
\item{path}{path to the folder containing MSDs or specific MSD file}

\item{type}{not required unless providing a folder in `path`;
default = "OU_IM_FY2*", other examples include: "PSNU_IM", "NAT_SUBNAT",
"PSNU", "Financial", "HRH"}
}
\description{
Extract MSD path on s3 from Data Mart
}
\keyword{internal}
