% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract MSD Meta Data}
\usage{
get_metadata(path, type, caption_note)
}
\arguments{
\item{path}{path to the folder containing MSDs or specific MSD file}

\item{type}{not required unless providing a folder in `path`; default = "OU_IM_FY19"
other examples include: "PSNU_IM", "NAT_SUBNAT", "PSNU", "Financial"}

\item{caption_note}{additional information to include in the footer}
}
\value{
vector of information related to what is being asked in `return`
}
\description{
This function is used to extract meta data as a list from the source file of
a MER Structured Dataset, MER NAT_SUBNAT Structured Dataset, Financial
Structure Dataset, or DATIM Genie export. It creates a list object, metadata,
in the global environment containing the source, current fiscal year, current
period, current quarter, as well as a caption.
}
\seealso{
Other metadata: 
\code{\link{extract_metadata}()},
\code{\link{source_info}()}
}
\concept{metadata}
