% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_pdap_loc}
\alias{get_pdap_loc}
\title{Get location
Identify where a user is working - locally or on workbench}
\usage{
get_pdap_loc()
}
\description{
Get location
Identify where a user is working - locally or on workbench
}
\keyword{internal}
