% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{handle_parquet}
\alias{handle_parquet}
\title{Read a parquet file}
\usage{
handle_parquet(file)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}
}
\description{
Read a parquet file
}
\keyword{internal}
