% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{handle_psd_format}
\alias{handle_psd_format}
\title{Read a tsv and parquet file}
\usage{
handle_psd_format(file)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}
}
\description{
Read a tsv and parquet file
}
\keyword{internal}
