% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_psd.R
\name{identify_psd}
\alias{identify_psd}
\title{Identify PSD Data Stream and Level}
\usage{
identify_psd(df)
}
\arguments{
\item{df}{PSD dataframe}
}
\value{
character, PSD data stream and the level
}
\description{
Identify PSD Data Stream and Level
}
\examples{
\dontrun{
#read in file for use
 path <- "~/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt"
 df <- read_psd(path)
#convert to RDS and delete the original txt file
 identify_psd(df_psnu) }
}
