% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{locate_bucket}
\alias{locate_bucket}
\title{Local PDAP Bucket}
\usage{
locate_bucket(file)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}
}
\description{
Identify which bucket to source for a given file when on PEPFAR Posit
Workbench - S3_READ or S3_WRITE
}
\keyword{internal}
