% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{process_msd}
\alias{process_msd}
\title{Processing to handle MSD as a zip/txt file}
\usage{
process_msd(
  file,
  save_rds = FALSE,
  remove_txt = FALSE,
  convert_to_old_names = FALSE
)
}
\arguments{
\item{file}{enter the full path to the MSD/FSD file}

\item{save_rds}{save the Structured Dataset as an rds file, default = FALSE}

\item{remove_txt}{should the txt file be removed, default = FALSE}
}
\description{
Processing to handle MSD as a zip/txt file
}
\keyword{internal}
