% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{read_msd}
\alias{read_msd}
\title{Import PEPFAR Structured Datasets .txt into R and covert to .rds}
\usage{
read_msd(
  file,
  save_rds = FALSE,
  remove_txt = FALSE,
  convert_to_old_names = FALSE
)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}

\item{save_rds}{save the Structured Dataset as an rds file, default = FALSE}

\item{remove_txt}{should the txt file be removed, default = FALSE}

\item{convert_to_old_names}{replace FY22Q2 naming convention with old?
default = FALSE}
}
\description{
Deprecated. Use `read_psd` instead.
}
