% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{read_psd}
\alias{read_psd}
\title{Import PEPFAR Structured Datasets to R}
\usage{
read_psd(file, export_format = "none", remove_base_file = FALSE)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}

\item{export_format}{if desired, save the PSD in another compressed format,
either "rds" or "parquet", default = "none"}

\item{remove_base_file}{should original base file be removed if exporting in
another compressed format? default = FALSE}
}
\description{
`read_psd` imports a stored PEPFAR Structured Datasets (.zip, .txt, or
.parquet). The function will read in a MSD, Genie, Financial or
HRH PEPFAR dataset, ensuring the column types are correct. The user has the
ability to store the txt file as a rds, significantly saving storage space
on the computer (and can then remove the txt file after importing). Most of
USAID/OHA processes and analyses rely on the use of the MSD file being read
in via `read_psd`
}
\examples{

\dontrun{
#convert Q1 clean PSNU file from txt to Rds
#read in file for use
 path <- "~/Data/MER_Structured_Datasets_OU_IM_FY22-24_20240315_v2_1.zip"
 df_psnu <- read_psd(path) }

}
