% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{rename_psd}
\alias{rename_psd}
\title{Rename PSD file when importing}
\usage{
rename_psd(file, ext)
}
\arguments{
\item{file}{enter the full path to the PEPFAR structured dataset file}

\item{ext}{new file extention}
}
\description{
Rename PSD file when importing
}
\keyword{internal}
