% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_ind.R
\docType{data}
\name{snapshot_ind}
\alias{snapshot_ind}
\title{MER Snapshot indicators}
\format{
A list of all snapshot indicators
\describe{
  \item{snapshot_ind}{indicator names}
}
}
\usage{
data(snapshot_ind)
}
\description{
List of indicators in the MER that are snapshot indicators, eg TX_CURR,
as opposed to a cumulative one, like HTS_TST. Snapshot indicators are handled
differently for calculating cumulative values and target achievement.
}
\keyword{datasets}
