% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_save.R
\name{split_save}
\alias{split_save}
\title{Export a single dataset into multiple files by group}
\usage{
split_save(df, group_var, folderpath, filename_stub, include_date = FALSE)
}
\arguments{
\item{df}{dataframe to split}

\item{group_var}{grouping variable to split the dataset by, eg operatingunit, funding_agency}

\item{folderpath}{directory where you want to store the files}

\item{filename_stub}{generic stub for naming all the files}

\item{include_date}{include date after filenamestub? default = FALSE, eg "20180913"}
}
\description{
`split_save` breaks a dataset by the groups provided and then exports those
individual frames as separate csv files. This process can be useful when
working with data across multiple partners and then creating datasets to
be sent to each of them with their own data.
}
\examples{
\dontrun{
#create country specific files for TX_NEW
 df_mer \%>\%
  filter(indicator == "TX_NEW",
         standardizeddisaggregate == "Total Numerator") \%>\%
  split_save(operatingunit, "~/CountryFiles", "FY18Q3_TX")
}
}
