% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{convert_coltype}
\alias{convert_coltype}
\title{Convert to specified column types}
\usage{
convert_coltype(df)
}
\arguments{
\item{df}{data frame from read_msd()}
}
\description{
Convert to specified column types
}
\keyword{internal}
