% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{convert_names}
\alias{convert_names}
\title{Convert variables back to old names if desired}
\usage{
convert_names(df, keep_old_names = FALSE)
}
\arguments{
\item{df}{data frame from read_msd()}

\item{keep_old_names}{replace FY22Q2 naming convention with old? default = F}
}
\description{
With the new changes to names introduced in FY22Q2, a user may have a desire
to keep the old naming convention in order to run old code without errors. To
do so, the user would specify `keep_old_names = TRUE`. Alternatively, if a
user reads in an older MSD, this function will default switch those names to
the current standard.
}
\keyword{internal}
