% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{convert_names}
\alias{convert_names}
\title{Convert variables for alignment}
\usage{
convert_names(df, retain_genie_cols)
}
\arguments{
\item{df}{data frame from read_psd()}

\item{retain_genie_cols}{should Genie specific columns (`dataelementuid`,
`categoryoptioncombouid`, `approvallevel`, `approvalleveldescription`) be
retained in the output dataset? default = FALSE}
}
\description{
Convert variables for alignment
}
\keyword{internal}
