% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psd.R
\name{rename_psd}
\alias{rename_psd}
\title{Rename PSD file when importing}
\usage{
rename_psd(file)
}
\arguments{
\item{file}{enter the full path to the PEPFAR Structured Dataset file,
eg "~/ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt"}
}
\description{
Rename PSD file when importing
}
\keyword{internal}
